Move to install command
