+++ src/extension/chatSessions/vscode-node/copilotCloudSessionsProvider.ts
@@ -7,7 +7,7 @@ import { RemoteAgentJobPayload } from '@vscode/copilot-api';
 import MarkdownIt from 'markdown-it';
 import * as pathLib from 'path';
 import * as vscode from 'vscode';
-import { Uri } from 'vscode';
+import { l10n, Uri } from 'vscode';
 import { IExperimentationService } from '../../../lib/node/chatLibMain';
 import { IAuthenticationService } from '../../../platform/authentication/common/authentication';
 import { IVSCodeExtensionContext } from '../../../platform/extContext/common/extensionContext';
@@ -200,6 +200,7 @@ export class CopilotCloudSessionsProvider extends Disposable implements vscode.C
 		@IExperimentationService private readonly _experimentationService: IExperimentationService,
 	) {
 		super();
+		this.registerCommands();
 
 		// Background refresh
 		getRepoId(this._gitService).then(async repoIds => {
@@ -258,6 +259,41 @@ export class CopilotCloudSessionsProvider extends Disposable implements vscode.C
 			this._register(onDebouncedAuthRefresh(() => this.refresh()));
 			this.telemetry.sendTelemetryEvent('copilotCloudSessions.refreshInterval', { microsoft: true, github: false }, telemetryObj);
 		});
+	}
+
+	private registerCommands() {
+		const checkoutPullRequestReroute = async (sessionItemOrResource?: vscode.ChatSessionItem | vscode.Uri) => {
+			const resource = sessionItemOrResource instanceof vscode.Uri
+				? sessionItemOrResource
+				: sessionItemOrResource?.resource;
+
+			if (!resource) {
+				return;
+			}
+
+			const pullRequestNumber = SessionIdForPr.parsePullRequestNumber(resource);
+			if (!pullRequestNumber) {
+				return;
+			}
+			const repoIds = await getRepoId(this._gitService);
+			if (!repoIds || repoIds.length === 0) {
+				vscode.window.showErrorMessage(l10n.t('No active repository found to checkout pull request.'));
+				return;
+			}
+
+			const installLabel = l10n.t('Install and Checkout');
+			const result = await vscode.window.showInformationMessage(
+				l10n.t('The GitHub Pull Requests extension is required to checkout this PR. Would you like to install and checkout?'),
+				{ modal: true },
+				installLabel
+			);
+
+			if (result === installLabel) {
+				await vscode.commands.executeCommand('workbench.extensions.installExtension', 'github.vscode-pull-request-github', { enable: true });
+				await vscode.commands.executeCommand('pr.checkoutFromDescription', { owner: repoIds[0].org, repo: repoIds[0].repo, number: pullRequestNumber });
+			}
+		};
+		this._register(vscode.commands.registerCommand('github.copilot.chat.checkoutPullRequestReroute', checkoutPullRequestReroute));
 	}
 
 	private getRefreshIntervalTime(hasHistoricalSessions: boolean): number {
